IA/requirements.txt
# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.8.0
Pillow>=10.0.0

# Dataset Management
datasets>=2.14.0  # HuggingFace datasets
albumentations>=1.3.1  # Data augmentation
pycocotools>=2.0.7  # COCO format handling

# Utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0

# Model Deployment
onnx>=1.14.0
onnxruntime>=1.15.0

# Logging & Monitoring
wandb>=0.15.0  # Optional: for experiment tracking
tensorboard>=2.13.0
